Did I just rewrite the game? YEP!
Really, there's no better way around this as everything's a breaking
change. The goal is to extract the CommandList/Pipeline usage to a
separate abstraction (called an `IVeldridPipeline`).

Deferred renderer doesn't compile for now.

In the process, I've also prepared `DeferredRenderer` to
override `Renderer` rather than `IRenderer`. The reason is primarly the
GlobalUniformBuffer and its usage inside
`VeldridRenderer.DrawVertices()`. This method is still not handled
properly with these changes (it needs to exist inside `Renderer`
itself), but that will come shortly.